Update ctdproc_steps.md using method used at RSMAS
SBE data processing software is used to convert SBE43 volts to umol/kg units applying Seabird temperature, pressure, salinity, time lag and hysteresis corrections.  A linear correction is done against bottle oxygen data for all stations pooled or by groups of stations if there is a clear shift in sensor calibration. The square root of the sum of squared residuals is computed to provide a metric of the quality of the calibrated CTD O2 data.  Typically, a value < 3 umol/kg is obtained.